Update script 6 v3

+++ 0x00-hello_world/6-size.c
@@ -20,7 +20,8 @@ int main(void)
 	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
 	printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(a));
 	printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(l));
-	printf("Size of a long long int: %llu bytes(s)\n", (unsigned long long)sizeof(k));
+	printf("Size of a long long int: %lu bytes(s)\n",
+	       (unsigned long)sizeof(k));
 	printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
